renamed original CodeUtil.H/cpp file to eXstaticUtil.H/cpp. Renamed old namespace MultiFab_Manipulation to eXstatic_MFab_Utils. FerroXUtils folder now contains FerroX_MFab_Utils namesspace and FerroX specific utils.
