+++ Source/Input/BoundaryConditions/BoundaryConditions.H
@@ -1,4 +1,7 @@
-/* Contributors: Saurabh Sawant
+/*
+ * This file is taken from eXstatic.
+ *
+ * Contributor: Saurabh S. Sawant
  *
  */
 #ifndef BOUNDARY_CONDITIONS_H_
