+++ Source/Utils/Make.package
@@ -2,5 +2,6 @@ VPATH_LOCATIONS   += $(CODE_HOME)/Source/Utils
 INCLUDE_LOCATIONS   += $(CODE_HOME)/Source/Utils
 
 include $(CODE_HOME)/Source/Utils/ablastr/Make.package
+include $(CODE_HOME)/Source/Utils/eXstaticUtils/Make.package
 include $(CODE_HOME)/Source/Utils/FerroXUtils/Make.package
 include $(CODE_HOME)/Source/Utils/SelectWarpXUtils/Make.package
